requirements.txt
pandas==1.1.3
librosa==0.8.0
scipy==1.5.2
numpy~=1.18.5
matplotlib==3.3.2
tensorflow==2.3.1
Pillow~=7.2.0
scikit_learn==0.23.2
streamlit~=0.69.2
scikit-learn~=0.23.2